requirements.txt
streamlit
langchain_community
langchain_groq
langchain_core
langchain
dotenv
faiss-cpu
langchain
load_dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdflangchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
pandas
openai
langchain-groq
pymupdf
huggingface_hub
python-dotenv